requirements.txt
Django==1.8.13
django-activity-stream==0.6.2
django-appconf==1.0.1
django-awesome-avatar==1.1.3
django-bootstrap-calendar==0.1.0
django-bootstrap-form==3.2.1
django-datetime-widget==0.9.3
django-friendship==1.3.3
django-image-cropping==1.0.2
django-jquery==1.12.2
django-jquery-ui==1.11.4
django-localflavor==1.1
django-model-utils==2.5
django-notifications-hq==1.2
django-static-jquery==2.1.1
django-templated-email==0.4.9
easy-thumbnails==2.2.1
jsmin==2.2.1
jsonfield==1.0.3
MySQL-python==1.2.5
Pillow==3.3.0
pytz==2016.4
six==1.10.0
tzlocal==1.2.2
